requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.3
annotated-types==0.7.0
anyio==4.12.1
apache-airflow==3.1.7
apache-airflow-core==3.1.7
apache-airflow-providers-common-compat==1.13.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.32.0
apache-airflow-providers-postgres==6.6.0
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-standard==1.11.0
apache-airflow-task-sdk==1.1.7
argcomplete==3.6.3
asgiref==3.11.0
asyncpg==0.31.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.13.4
behave==1.2.6
cadwyn==5.4.6
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==42.0.8
cssselect==1.3.0
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
execnet==2.1.1
fastapi==0.117.1
fastapi-cli==0.0.20
fasteners==0.20
filelock==3.25.0
fsspec==2026.1.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenback==1.3.0
greenlet==3.3.1
grpcio==1.65.5
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==6.0.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.20.0
mycdp==1.3.2
natsort==8.4.0
numpy==2.4.2
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.11.7
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
parameterized==0.9.0
parse==1.21.1
parse_type==0.6.6
pathspec==1.0.4
pdbp==1.8.2
pendulum==3.2.0
platformdirs==4.9.2
pluggy==1.6.0
protobuf==4.25.8
psutil==7.2.2
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.11.0
pynose==1.5.5
pyotp==2.9.0
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.1
pytest-xdist==3.6.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-slugify==8.0.4
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
rich==14.3.3
rich-argparse==1.7.2
rich-toolkit==0.18.1
rpds-py==0.30.0
sbvirtualdisplay==1.4.0
selenium==4.32.0
seleniumbase==4.38.3
setproctitle==1.3.7
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soccerdata==1.8.8
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.46
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
tabcompleter==1.4.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.3.0
text-unidecode==1.3
tqdm==4.67.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
Unidecode==1.4.0
universal_pathlib==0.2.6
urllib3==1.26.20
uuid6==2025.0.1
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.46.3
wirerope==1.0.0
wrapper-tls-requests==1.2.5
wrapt==2.1.1
wsproto==1.2.0
zipp==3.23.0
